add cache force disable property

+++ ci-operator/step-registry/ipi/conf/vsphere/ipi-conf-vsphere-commands.sh
@@ -111,27 +111,37 @@ networking:
   machineNetwork:
   - cidr: "${machine_cidr}"
 EOF
+PULL_THROUGH_CACHE_DISABLE="/var/run/vault/vsphere-config/pull-through-cache-disable"
 
-if [ ${PULL_THROUGH_CACHE} == "enabled" ]; then
-  echo "$(date -u --rfc-3339=seconds) - pull-through cache enabled for job"
-  PULL_THROUGH_CACHE_CREDS="/var/run/vault/vsphere-config/pull-through-cache-secret"
-  PULL_THROUGH_CACHE_CONFIG="/var/run/vault/vsphere-config/pull-through-cache-config"
-  PULL_SECRET="/var/run/secrets/ci.openshift.io/cluster-profile/pull-secret"
-  TMP_INSTALL_CONFIG="/tmp/tmp-install-config.yaml"
-  if [ -f ${PULL_THROUGH_CACHE_CREDS} ]; then    
-    echo "$(date -u --rfc-3339=seconds) - pull-through cache credentials found. updating pullSecret"
-    cat ${CONFIG} | sed '/pullSecret/d' > ${TMP_INSTALL_CONFIG}2
-    cat ${TMP_INSTALL_CONFIG}2 | sed '/\"auths\"/d' > ${TMP_INSTALL_CONFIG}
-    jq -cs '.[0] * .[1]' ${PULL_SECRET} ${PULL_THROUGH_CACHE_CREDS} > /tmp/ps-combined.json
-    echo -e "\npullSecret: '""$(cat /tmp/ps-combined.json)""'" >> ${TMP_INSTALL_CONFIG}
-    cat ${TMP_INSTALL_CONFIG} > ${CONFIG}
-  else
-    echo "$(date -u --rfc-3339=seconds) - pull-through cache credentials not found. not updating pullSecret"
+CACHE_FORCE_DISABLE="false"
+if [ -f "${PULL_THROUGH_CACHE_DISABLE}" ]; then
+  CACHE_FORCE_DISABLE=$(cat ${PULL_THROUGH_CACHE_DISABLE})
+fi
+
+if [ ${CACHE_FORCE_DISABLE} == "false" ]; then
+  if [ ${PULL_THROUGH_CACHE} == "enabled" ]; then
+    echo "$(date -u --rfc-3339=seconds) - pull-through cache enabled for job"
+    PULL_THROUGH_CACHE_CREDS="/var/run/vault/vsphere-config/pull-through-cache-secret"
+    PULL_THROUGH_CACHE_CONFIG="/var/run/vault/vsphere-config/pull-through-cache-config"
+    PULL_SECRET="/var/run/secrets/ci.openshift.io/cluster-profile/pull-secret"
+    TMP_INSTALL_CONFIG="/tmp/tmp-install-config.yaml"
+    if [ -f ${PULL_THROUGH_CACHE_CREDS} ]; then    
+      echo "$(date -u --rfc-3339=seconds) - pull-through cache credentials found. updating pullSecret"
+      cat ${CONFIG} | sed '/pullSecret/d' > ${TMP_INSTALL_CONFIG}2
+      cat ${TMP_INSTALL_CONFIG}2 | sed '/\"auths\"/d' > ${TMP_INSTALL_CONFIG}
+      jq -cs '.[0] * .[1]' ${PULL_SECRET} ${PULL_THROUGH_CACHE_CREDS} > /tmp/ps-combined.json
+      echo -e "\npullSecret: '""$(cat /tmp/ps-combined.json)""'" >> ${TMP_INSTALL_CONFIG}
+      cat ${TMP_INSTALL_CONFIG} > ${CONFIG}
+    else
+      echo "$(date -u --rfc-3339=seconds) - pull-through cache credentials not found. not updating pullSecret"
+    fi
+    if [ -f ${PULL_THROUGH_CACHE_CONFIG} ]; then
+      echo "$(date -u --rfc-3339=seconds) - pull-through cache configuration found. updating install-config"
+      cat ${PULL_THROUGH_CACHE_CONFIG} >> ${CONFIG}
+    else
+      echo "$(date -u --rfc-3339=seconds) - pull-through cache configuration not found. not updating install-config"
+    fi
   fi
-  if [ -f ${PULL_THROUGH_CACHE_CONFIG} ]; then
-    echo "$(date -u --rfc-3339=seconds) - pull-through cache configuration found. updating install-config"
-    cat ${PULL_THROUGH_CACHE_CONFIG} >> ${CONFIG}
-  else
-    echo "$(date -u --rfc-3339=seconds) - pull-through cache configuration not found. not updating install-config"
-  fi
+else 
+  echo "$(date -u --rfc-3339=seconds) - pull-through cache force disabled"
 fi
